feat:api 점검

+++ src/main/java/org/sopt/dto/review/response/EmotionStatisticDTO.java
@@ -1,5 +1,5 @@
 package org.sopt.dto.review.response;
 
 public record EmotionStatisticDTO(String emotionTag,
-                                  Float percentage) {
+                                  int percentage) {
 }
