+++ src/main/java/org/sopt/dto/review/response/StarStatisticDTO.java
@@ -1,5 +1,5 @@
 package org.sopt.dto.review.response;
 
-public record StarStatisticDTO(Long starValue,
-                               Float percentage) {
+public record StarStatisticDTO(int starValue,
+                               int percentage) {
 }
